cloud_functions/kmeans-model/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.4.2
google-cloud-storage==2.5.0
pandas==1.3.3
numpy==1.21.6
scikit-learn==1.0.2
joblib==1.1.0
db-dtypes